[utils] Deprecate `jwt_encode_hs256`
Authored by: bashonly
